retain entry position when printing tables
AN ALTERNATIVE APPROACH HAS ALREADY BEEN MERGED.
SAVING IN CASE WE WANT TO GO WITH THIS INSTEAD.

tabulated-list-entries are now uniquely indexed by a numeric id.
This should prevent the cursor from jumping to the top of the buffer
while the log is being populated.

related: https://github.com/progfolio/elpaca/issues/33